Only retry-on-reject if iterators are paused
Due to recent commits, iterators are not paused if target nodes are 
overloaded. With enabled retries this can lead of having many request
inside the memory and worst to a OOM. To avoid that, we disable the 
retry logic and let a possible rejected execution exception bubble up
to the client.

Followup of 6ab7b340f248d34f1ea8a0bf9a5905222b0ea247 and
182d471ddea4ad764c2203a610cbaca8ee648a00.